requirements.txt
asgiref==3.7.2
Babel==2.12.1
boto3==1.28.44
botocore==1.31.44
certifi==2023.7.22
charset-normalizer==3.2.0
coinbase-commerce==1.0.1
dj-database-url==2.1.0
Django==4.2.3
django-countries==7.5.1
django-phonenumber-field==7.1.0
django-shortuuidfield==0.1.3
django-storages==1.14
gunicorn==21.2.0
idna==3.4
jmespath==1.0.1
packaging==23.1
phonenumbers==8.13.17
Pillow==10.0.0
psycopg2-binary==2.9.7
python-dateutil==2.8.2  
python-decouple==3.8
requests==2.31.0
s3transfer==0.6.2
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
white==0.1.2
